+++ src/flatObject/index.js
@@ -17,12 +17,20 @@ function flatObject ( dependencies, d ) {
             , [ flatIO, flatStore ]   = flatData ( dependencies, d )
             , objectAPI = {
                     // I/O Operations
-                      insert : insert ( dependencies, flatIO )                // Extends available data;
+                      insertSegment : insert ( dependencies, flatIO )         // Extends available data with new data segment;
                     , export : ex     ( dependencies, flatIO )                // Returns flat data;
                     , copy   : copy   ( dependencies, flatIO )                // Creates deep copy of original data
                     , model  : model  ( dependencies, flatIO, flatStore )      // Arrange data according specific data-model. Model should come as a function;
                     , query  : query  ( dependencies, flatIO, flatStore )      // Request, and evaluate data. Returns a new flat object;
                     , setupFilter : setupFilter ( flatIO )                    // Functions should filter content according some criteria. Generated indexes will help for data search in
+                    , listSegments  : () =>  {
+                                    const ix = Object.keys ( flatIO.getIndexes() );
+                                    return ix.reduce ( (res,item) => {
+                                                              if ( item.includes('/') )   return res
+                                                              res.push ( item )
+                                                              return res
+                                                        }, [] )
+                                }
                     , index  : n => {
                                         if ( n == null )   return null
                                         let r  = flatIO.getLine(n)