+++ src/flatObject/insert.js
@@ -2,12 +2,22 @@
 
 function insert ( dependencies, flatIO ) {
 return function insert ( name, inData ) {
-            
-            if ( !inData.export && !(typeof( inData.export) === 'function'))   return null
-            let 
-                  d    = inData.export ()
-                , copy = inData.copy ()
-                ;
+            const { convert, isDTO, isDTM, walk } = dependencies();
+            let d, copy;
+
+            if ( isDTO(inData) ) {
+                        d    = inData.export ()
+                        copy = inData.copy ()
+                        
+               }
+            else if ( isDTM(inData) ) {
+                        d = walk ({ data : inData })
+                        copy = walk ({ data : inData })
+                }
+            else {
+                        [ copy,, d ] = convert.from ( 'std' ).toFlat ( dependencies, inData )
+                        console.warn ( 'A non "dt-object" data segment was inserted. Autoconverted to "dt-object".' )
+                }
             
             const search = new RegExp ( `^root\/` );
             d.forEach ( line => {   // Word 'root' should be changed to the argument 'name'.